Add (possibly outdated) support for Arch Linux
Thanks Tomasz Żok (@tzok) for the long-ignored PR #1 which added this.
I just stole the LIBC/LIBC32 paths from that, rather than the full
changes with the bashisms and the updated headers (which would be
outdated again).

I confess to not having tested this on Arch Linux — I suspect the glibc
2.33 issues will show up, making it useless — but do promise to actually
actually changes to fix it if broken.