Tentative glibc 2.33 stat function support
The stat() family of functions moved from using an older, #define based
system of compatibility to symbols versions in glibc 2.33, which breaks
bingcc horribly. This is because the new headers link to actual stat,
fstat, etc, functions but older glibc versions do not have such
functions at all: instead there are __xstat, __fxstat, etc functions
which accept the version as an argument.

There's no easy way of fixing this with just headers, so add a new
object which implements the new function signatures in terms of the old
ones. This glibc_wrapper.o is then linked wherever possible (spitting
out lots of warnings).

gensymoverride is updated to read a list of symbols from
forbidden_symbols and not write any version overrides for them, so that
we link to our new wrappers correctly.